backend/ml-models/requirements.txt
# ML Training Requirements
torch==2.1.0
torchvision==0.16.0
tensorflow==2.14.0
numpy>=1.26.0,<2.0.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
Pillow==10.1.0

# Satellite data
rasterio==1.3.9
sentinelsat==1.2.1
netCDF4==1.6.5
h5py==3.10.0

# Data processing
tqdm==4.66.1
requests==2.31.0

# Ice data sources
xarray==2023.12.0
